Add validators for profile and image
